+++ src/Profiles/ProfileInterface.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * @copyright   ©2025 Maatify.dev
+ * @Liberary    maatify/data-adapters
+ * @Project     maatify:data-adapters
+ * @author      Mohamed Abdulalim (megyptm) <mohamed@maatify.dev>
+ * @since       2025-11-14 15:26
+ * @see         https://www.maatify.dev Maatify.com
+ * @link        https://github.com/Maatify/data-adapters  view project on GitHub
+ * @note        Distributed in the hope that it will be useful - WITHOUT WARRANTY.
+ */
+
+declare(strict_types=1);
+
+
+namespace Maatify\DataAdapters\Profiles;
+
+interface ProfileInterface
+{
+    public function value(): string;
+}